Restructure market submodule data classes
