+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/market/top/MarketListTopDataSource.kt
@@ -1,6 +1,7 @@
 package io.horizontalsystems.bankwallet.modules.market.top
 
 import io.horizontalsystems.bankwallet.core.IRateManager
+import io.horizontalsystems.bankwallet.modules.market.SortingField
 import io.horizontalsystems.xrateskit.entities.CoinMarket
 import io.reactivex.Observable
 import io.reactivex.Single
